remove duplicate weighted columns from tiger

+++ tasks/meta.py
@@ -139,7 +139,7 @@ class OBSColumn(Base):
     description = Column(Text) # human-readable description to provide in
                                  # bigmetadata
 
-    weight = Column(Integer, default=0)
+    weight = Column(Numeric, default=0)
     aggregate = Column(Text) # what aggregate operation to use when adding
                                # these together across geoms: AVG, SUM etc.
